[fix] ファイルの場所を整理した
