+++ docker-compose.yml
@@ -3,7 +3,7 @@ services:
   go-gin:
     container_name: go-gin
     hostname: go-gin
-		build: ./back/app
+    build: ./back/app
     restart: always
     depends_on:
       - mysql
